dev: add font + some else
