+++ front/src/utils/axios.ts
@@ -5,6 +5,21 @@ export const myRequest = axios.create({
     timeout: 1000
 });
 
+myRequest.interceptors.request.use(
+    function (config) {
+        const token = localStorage.getItem('token');
+        config.headers = {
+            ...config.headers,
+            'Authorization': `Bearer ${token}`
+        }
+        return config;
+    },
+    function (error) {
+        // Do something with request error
+        return Promise.reject(error);
+    }
+);
+
 myRequest.interceptors.response.use(
     (response) => {
         return response.data;
